AuthenticationElytronTestCase - EJB/remoting tests with Elytron
