+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/security/users.properties
@@ -1,3 +1,4 @@
+#$REALM_NAME=UsersRoles$
 user1=password1
 user2=password2
 admin=admin
